Initial commit.
'lumo_calendar_weekview' creates a class for viewing weeks.
This is the first rough draft of the class which has a lot of
hardcoded data.

It doesn't represent the finished data flow, but running it shows
how the basic layouts will work.

Basically it uses a function 'block_zipper' to print two days
at a time, side by side. Of course, it then leaves the task
to properly order the pairing of blocks so that when printed
the week appears in order.