Complete paginate view for weeks.
Update '_roll_forward' and '_roll_backward' to work when in Day mode and in Week mode.

Introduce 'times_formatter' so that week view always displays properly. (Previously nothing was aligning properly)
Create functions for 'military_to_standard' and 'standard_to_military'.
Update 'list_safe_get' to return values that match the same format as the events that are in DayBlock.events.
Update 'get_header_date' for CalendarPageWeek so that it always shows the right month, even for weeks that split a month.
(This is based on counting which month is more represented in the given 7 days.)

Change double quotes to single quotes.