Dirty(ish) commit.
Solve the issue with programs recursively running the 'root loop' causing 'quit' to be required for each time a sub-program was run.
Turn menus into objects to be able to have persistent menus even when input isn't recognized; i.e. "Reloop"
Refactor router if/elif routes to use dictionaries so that the shortcut letters for the user can change but they alwyas map to values in the dict that are
global for the router. The router checks if either the user has entered the full word, or has used a shortcut, which looks up the full word.

Make pomodoro integrate with __main__ properly. It's now flexible to accept minutes from the command line or ask user for input.

Update lumocards-one.txt to match current TODO.