Feature: get confirmation menu in update_event_1() to be correctly formatted.
Bug fix: Check for location and description as literal 'none' in lumo_calendar_actions': create_event().
Previously it was using dict_event.get() which meant that it always uploaded a None type to google event. This mirrors google
in that if there is no location, decription, etc. nothing is added to the dict.

Feature: add simple, basic feedback to each of the routes in the confirmation menu in _event_actions_router() and
same thing to each of the routes in update_event_1().

Feature: add new function display_menu_confirmation() to 'lumo_calendar_utils'.