Remove non-essential features / menu items.
Strip pomodoro to not use the logging function and other similar functions.
Create display_2() which allows menu to show the amt of minutes going into the
next 'focus_round'.

Create a new timer class called TimerDev which allows for easy testing
because it only prints the timer values without sleeping.